canvas: move to separate CGI script
* stream: and passFd is now client-based, and accessible for buffers
* Bitmap's width & height is now int, not uint64
* no more non-network Bitmap special case in the pager for canvas

I just shoehorned it into the static image model, so it still doesn't
render changes after page load. But at least now it doesn't crash the
browser.

+++ src/img/bitmap.nim
@@ -3,8 +3,8 @@ import types/color
 type
   Bitmap* = ref object of RootObj
     px*: seq[RGBAColorBE]
-    width*: uint64
-    height*: uint64
+    width*: int
+    height*: int
 
   ImageBitmap* = ref object of Bitmap
 
@@ -13,27 +13,27 @@ type
     imageId*: int
     contentType*: string
 
-proc newBitmap*(width, height: uint64): ImageBitmap =
+proc newBitmap*(width, height: int): ImageBitmap =
   return ImageBitmap(
     px: newSeq[RGBAColorBE](width * height),
     width: width,
     height: height
   )
 
-proc setpx*(bmp: Bitmap; x, y: uint64; color: RGBAColorBE) {.inline.} =
+proc setpx*(bmp: Bitmap; x, y: int; color: RGBAColorBE) {.inline.} =
   bmp.px[bmp.width * y + x] = color
 
-proc setpx*(bmp: Bitmap; x, y: uint64; color: ARGBColor) {.inline.} =
+proc setpx*(bmp: Bitmap; x, y: int; color: ARGBColor) {.inline.} =
   bmp.px[bmp.width * y + x] = rgba_be(color.r, color.g, color.b, color.a)
 
-proc getpx*(bmp: Bitmap; x, y: uint64): RGBAColorBE {.inline.} =
+proc getpx*(bmp: Bitmap; x, y: int): RGBAColorBE {.inline.} =
   return bmp.px[bmp.width * y + x]
 
-proc setpxb*(bmp: Bitmap; x, y: uint64; c: RGBAColorBE) {.inline.} =
+proc setpxb*(bmp: Bitmap; x, y: int; c: RGBAColorBE) {.inline.} =
   if c.a == 255:
     bmp.setpx(x, y, c)
   else:
     bmp.setpx(x, y, bmp.getpx(x, y).blend(c))
 
-proc setpxb*(bmp: Bitmap; x, y: uint64; c: ARGBColor) {.inline.} =
+proc setpxb*(bmp: Bitmap; x, y: int; c: ARGBColor) {.inline.} =
   bmp.setpxb(x, y, rgba_be(c.r, c.g, c.b, c.a))
